+++ programs/drift/src/instructions/keeper.rs
@@ -806,6 +806,7 @@ pub fn handle_resolve_perp_pnl_deficit(
 
         // reload the spot market vault balance so it's up-to-date
         ctx.accounts.spot_market_vault.reload()?;
+        ctx.accounts.insurance_fund_vault.reload()?;
         math::spot_withdraw::validate_spot_market_vault_amount(
             spot_market,
             ctx.accounts.spot_market_vault.amount,
@@ -934,6 +935,7 @@ pub fn handle_resolve_perp_bankruptcy(
 
         // reload the spot market vault balance so it's up-to-date
         ctx.accounts.spot_market_vault.reload()?;
+        ctx.accounts.insurance_fund_vault.reload()?;
         math::spot_withdraw::validate_spot_market_vault_amount(
             spot_market,
             ctx.accounts.spot_market_vault.amount,
@@ -1039,6 +1041,7 @@ pub fn handle_resolve_spot_bankruptcy(
 
         // reload the spot market vault balance so it's up-to-date
         ctx.accounts.spot_market_vault.reload()?;
+        ctx.accounts.insurance_fund_vault.reload()?;
         math::spot_withdraw::validate_spot_market_vault_amount(
             spot_market,
             ctx.accounts.spot_market_vault.amount,